+++ src/main/java/com/side/portfolio/demo/dto/CartForm.java
@@ -3,11 +3,13 @@ package com.side.portfolio.demo.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter @Setter
 public class CartForm {
 
     private Long teamId;
     private Long itemId;
     private int qty;
-    private float price;
+    private BigDecimal price;
 }
